feat(auction): InterHandler real implementation + auction item data headers (DnfItemInfo/ROI_Category/RandomOption etc.); fix IArea/GSArea vtables, ITimeEntity base layout, TE_Entity regist/ctor/operator(), ISession inline weaks, GameDataPool namespace/const; MISSING 3090→3052

+++ dnf_decompile/source/DNFServer/GameServer/ServerLab/Auction/AuctionServer/auction_source/HandlerFor_TE_.h
@@ -15,12 +15,12 @@ public:
     virtual ~HandlerFor_TE_();
     virtual void init();
     void initTimeEvent();
-    unsigned int onTIME_AUCTION_EXPIRE_EVENT_CHECK(nsl::InternalMsg* pArg);
-    unsigned int onTIME_AUCTION_STATISTICS_COLLECTOR(nsl::InternalMsg* pArg);
-    unsigned int onTIME_CHECK_CONFIG(nsl::InternalMsg* pArg);
-    unsigned int onTIME_AUCTION_DB_PING(nsl::InternalMsg* pArg);
-    unsigned int onTIME_AUCTION_TRY_SHUTDOWN(nsl::InternalMsg* pArg);
-    unsigned int onTIME_AUCTION_UPDATE_AVERAGE_PRICE(nsl::InternalMsg* pArg);
+    unsigned long onTIME_AUCTION_EXPIRE_EVENT_CHECK(nsl::InternalMsg* pArg);
+    unsigned long onTIME_AUCTION_STATISTICS_COLLECTOR(nsl::InternalMsg* pArg);
+    unsigned long onTIME_CHECK_CONFIG(nsl::InternalMsg* pArg);
+    unsigned long onTIME_AUCTION_DB_PING(nsl::InternalMsg* pArg);
+    unsigned long onTIME_AUCTION_TRY_SHUTDOWN(nsl::InternalMsg* pArg);
+    unsigned long onTIME_AUCTION_UPDATE_AVERAGE_PRICE(nsl::InternalMsg* pArg);
 
     struct tm mOldTM;
     nsl::TE_Entity<HandlerFor_TE_>* mTimeEntity[125];
